131 - solve circular dependency issue
