+++ tests/tel2puml/pv_to_puml/walk_puml_graph/test_walk_puml_logic_graph.py
@@ -3,8 +3,9 @@
 import networkx as nx
 
 from tel2puml.pv_to_puml.walk_puml_graph.node import Node
-from tel2puml.utils_test import (
-    merge_markov_without_loops_and_logic_detection_analysis
+from .utils import (
+    merge_markov_without_loops_and_logic_detection_analysis,
+    get_event_reference_from_events
 )
 from tel2puml.pv_to_puml.walk_puml_graph.walk_puml_logic_graph import (
     create_puml_graph_from_node_class_graph,
@@ -20,7 +21,6 @@ from tel2puml.pv_to_puml.data_ingestion import (
     update_all_connections_from_clustered_events,
 )
 from tel2puml.events import events_to_markov_graph
-from tel2puml.utils_test import get_event_reference_from_events
 from tel2puml.pv_event_simulator import (
     generate_test_data_event_sequences_from_puml,
 )
